nu med fields data

+++ src/BankAccount.java
@@ -37,7 +37,7 @@ public class BankAccount{
     }
 
 
-    public float getBalance() {
+    public int getBalance() {
         this.stalkingCounter++;// Den her linje skal demonstrere en af de ting jeg som programmør kan få ud af begrænse adgangen klassens felter (bemærk feltet 'balance' i toppen af klassen er private).
                                // Hver gang instansens balance attribut skal 'ses' udefra, skal det ske gennem denne metode og jeg kan så tilføje lidt  kode der kører hver gang dette sker
                               // I dette eksempel bliver en 'stalkingCounter' talt op. M.a.o.: jeg tracker data om udefrakommende interesse i feltet saldo.
